requirements.txt
# Core dependencies
streamlit>=1.28.0
snowflake-connector-python>=3.3.0
snowflake-snowpark-python>=1.8.0
pandas>=2.0.0
plotly>=5.15.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
loguru>=0.7.0
pyyaml>=6.0
python-dotenv>=1.0.0
pyarrow<19.0.0

# Development and testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0

# Documentation dependencies
mkdocs>=1.5.0
mkdocs-material>=9.2.0
mkdocstrings[python]>=0.22.0 